[#7] feat : add value object class label
- vo table인 label 생성후 issue와 연결 했습니다.
