--- BE/issue-tracker/src/main/java/dev/codesquad/issuetracker/domain/Milestone.java
@@ -1,40 +1,0 @@
-package dev.codesquad.issuetracker.domain;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Entity
-@Getter
-@ToString
-@NoArgsConstructor
-public class Milestone {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotNull
-    private String title;
-
-    @NotNull
-    private String description;
-
-    @NotNull
-    private boolean status;
-
-    @NotNull
-    private LocalDate dueDate;
-
-    @OneToMany(mappedBy = "milestone")
-    private List<Issue> issues = new ArrayList<>();
-}
