+++ BE/issue-tracker/src/main/java/dev/codesquad/issuetracker/domain/milestone/Milestone.java
@@ -1,13 +1,13 @@
-package dev.codesquad.issuetracker.domain;
+package dev.codesquad.issuetracker.domain.milestone;
 
+import dev.codesquad.issuetracker.domain.issue.Issue;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToMany;
 import javax.persistence.OneToMany;
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
@@ -18,23 +18,24 @@ import lombok.ToString;
 @Getter
 @ToString
 @NoArgsConstructor
-public class User {
+public class Milestone {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private String githubId;
+    private String title;
 
-    private String profileUrl;
+    @NotNull
+    private String description;
 
-    @OneToMany(mappedBy = "user")
-    private List<Comment> comments = new ArrayList<>();
+    @NotNull
+    private boolean status;
 
-    @ManyToMany(mappedBy = "users")
+    @NotNull
+    private LocalDate dueDate;
+
+    @OneToMany(mappedBy = "milestone")
     private List<Issue> issues = new ArrayList<>();
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<Issue> issueList = new ArrayList<>();
 }
