+++ BE/issue-tracker/src/main/java/dev/codesquad/issuetracker/domain/user/User.java
@@ -1,0 +1,42 @@
+package dev.codesquad.issuetracker.domain.user;
+
+import dev.codesquad.issuetracker.domain.issue.Comment;
+import dev.codesquad.issuetracker.domain.issue.Issue;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import javax.persistence.OneToMany;
+import javax.validation.constraints.NotNull;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Getter
+@ToString
+@NoArgsConstructor
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotNull
+    private String githubId;
+
+    private String profileUrl;
+
+    @OneToMany(mappedBy = "user")
+    private List<Comment> comments = new ArrayList<>();
+
+    @ManyToMany(mappedBy = "users")
+    private List<Issue> issues = new ArrayList<>();
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Issue> issueList = new ArrayList<>();
+}
