+++ BE/issue-tracker/src/main/java/dev/codesquad/issuetracker/domain/vo/Label.java
@@ -1,0 +1,17 @@
+package dev.codesquad.issuetracker.domain.vo;
+
+import javax.persistence.Embeddable;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Embeddable
+@Getter
+@ToString
+@NoArgsConstructor
+public class Label {
+
+    private String name;
+    private String description;
+    private String colorCode;
+}
